Remove automatic dispatcher selection
Previously we added a feature that automatically selects the dispatcher that contains the coroutine that we have placed breakpoint in. Unfortunately, to work this feature iterated through the first frames of all the coroutines, which under the hood resulted in unwinding the whole stack traces of all the coroutines in the application. This dramatically slowed down the debugging process, because we effectively captured the complete coroutine dump on each stepping and breakpoint hit event. This commit removes this feature to speed up the panel building time.